# src/main/resources/openapi.yaml
openapi: 3.0.3

info:
  title: Order Service API
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:

  /orders:
    post:
      tags:
        - Order
      summary: Create order
      operationId: createOrder

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'

      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'

    get:
      tags:
        - Order
      summary: Get all orders of the authenticated user
      operationId: getOrders

      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OrderStatusApi'

      responses:
        '200':
          description: Orders found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetOrderResponse'

  /orders/{orderId}:
    get:
      tags:
        - Order
      summary: Get order by id
      operationId: getOrderById

      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid

      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetOrderResponse'

        '404':
          description: Order not found

components:

  schemas:

    CreateOrderRequest:
      type: object

      required:
        - items

      properties:
        items:
          type: array
          minItems: 1

          items:
            $ref: '#/components/schemas/OrderItem'

    OrderItem:
      type: object

      required:
        - productId
        - quantity

      properties:
        productId:
          type: string
          format: uuid

        quantity:
          type: integer
          minimum: 1

    CreateOrderResponse:
      type: object

      properties:
        orderId:
          type: string

        status:
          $ref: '#/components/schemas/OrderStatusApi'

    GetOrderResponse:
      type: object

      properties:
        orderId:
          type: string

        customerId:
          type: string
          format: uuid

        status:
          $ref: '#/components/schemas/OrderStatusApi'

        createdAt:
          type: string
          format: date-time

        items:
          type: array

          items:
            $ref: '#/components/schemas/OrderItem'

    OrderStatusApi:
      type: string

      enum:
        - CREATED
        - PAYMENT_PENDING
        - PAYMENT_APRROVED
        - COMPLETED